+++ System/AutoCloseDoors.cs
@@ -73,5 +73,6 @@ namespace AutoCloseDoors.Systems
                 }
             }
         }
+
     }
 }